Rename registrationserver to server (#18)
* renaming registrationserver to server

* fixing tests
